// apps/admin/src/components/layout/types.ts
import { LinkProps } from "@tanstack/react-router";

export interface User {
    name: string;
    email: string;
    avatar: string;
}

export interface Application {
    name: string;
    logo: React.ElementType;
}

interface BaseNavItem {
    title: string;
    badge?: string;
    icon?: React.ElementType;
}

type NavLink = BaseNavItem & {
    url: LinkProps["to"];
    items?: never;
};

type NavCollapsible = BaseNavItem & {
    items: (BaseNavItem & { url: LinkProps["to"] })[];
    url?: never;
};

type NavItem = NavCollapsible | NavLink;

interface NavGroup {
    title: string;
    items: NavItem[];
}

interface SidebarData {
    navGroups: NavGroup[];
}

export type { SidebarData, NavGroup, NavItem, NavCollapsible, NavLink };